requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
PySocks==1.7.1
requests==2.32.3
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
telegram==0.0.1
playwright~=1.47.0
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
python-telegram-bot==21.6